+++ modules/@apostrophecms/ui/ui/apos/mixins/AposThemeMixin.js
@@ -1,12 +1,11 @@
+import { _themeClass } from '../composables/AposTheme.js';
+
 // Provides computed classes for decorating top-level Apos vue apps with a UI
 // theme
-
 export default {
   computed: {
     themeClass() {
-      const classes = [];
-      classes.push(`apos-theme--primary-${window.apos.ui.theme.primary}`);
-      return classes;
+      return _themeClass();
     }
   }
 };